encriptaçao e autorização

+++ routes/tasks.js
@@ -1,14 +1,10 @@
 module.exports = app =>{
     const Tasks = app.db.models.Tasks;
     app.route("/tasks")
-    // .all((req, res, next) => {
-    //     // middleware de pré-execução
-    //     delete req.body.id;
-    //     next();
-    // })  
+    .all(app.auth.authenticate())  
     .get((req, res) => {
         //retorna todas as tarefas
-        Tasks.findAll({})
+        Tasks.findAll({where: {user_id: req.user_id}})
             .then(result => res.json(result))
             .catch(error => {
                 res.status(412).json({msg: error.message});
@@ -16,6 +12,7 @@ module.exports = app =>{
     })
     .post((req, res) => {
         // cadastra uma nova tarefa
+        req.body.user_id = req.user_id;
         Tasks.create(req.body)
             .then(result => res.json(result))
             .catch(error => {
@@ -23,14 +20,10 @@ module.exports = app =>{
             });
     });
     app.route("/tasks/:id")
-    // .all((req, res, next) => {
-    //     //middleware de pré-execução
-    //     delete req.body.id;
-    //     next();
-    // })
+    .all(app.auth.authenticate())  
     .get((req, res) => {
         // consulta tarefa
-        Tasks.findOne({where: req.params})
+        Tasks.findOne({where: {id: req.params.id, user_id: req.user_id}})
             .then(result => {
                 if(result){
                     res.json(result);
@@ -44,7 +37,7 @@ module.exports = app =>{
     })
     .put((req, res) => {
         //atualiza tarefa
-        Tasks.update(req.body, {where: req.params})
+        Tasks.update(req.body, {where: {id: req.params.id, user_id: req.user_id}})
             .then(result => res.sendStatus(204))
             .catch(error => {
                 res.status(412).json({msg: error.message})
@@ -52,7 +45,7 @@ module.exports = app =>{
     })
     .delete((req, res) => {
         //deleta tarefa
-        Tasks.destroy({where: req.params})
+        Tasks.destroy({where: {id: req.params.id, user_id: req.user_id}})
             .then(result => {
                 res.sendStatus(204);
             })
